pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ezauth"
version = "0.1.0"
description = "Multi-tenant authentication service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "boto3>=1.35",
    "chevron>=0.14",
    "premailer>=3.10",
    "python-jose[cryptography]>=3.3",
    "argon2-cffi>=23.1",
    "loguru>=0.7",
    "httpx>=0.28",
    "aiodns>=3.2",
    "jinja2>=3.1",
    "python-multipart>=0.0.18",
    "cryptography>=44.0",
    "email-validator>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "fakeredis>=2.26",
    "factory-boy>=3.3",
    "time-machine>=2.16",
    "ruff>=0.8",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ezauth"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
